+++ src/app/tasks/page.tsx
@@ -1,73 +1,16 @@
 "use client";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faChevronDown } from "@fortawesome/free-solid-svg-icons";
-import { IoArrowForwardOutline } from "react-icons/io5";
 import styles from "./page.module.scss";
-import React, { useEffect, useState, useContext } from "react";
+import React, { useEffect, useState} from "react";
 import { fetchTasks } from "@/services/api";
-import dayjs from "dayjs";
-import { ProjectsContext } from "@/contexts/projects";
 import { PageInfo, Task } from "@/services/api";
-
-const statusMap: Record<string, string> = {
-  scheduled: "未完了",
-  completed: "完了",
-};
-const getStatusLabel = (status: string) => statusMap[status] || "不明";
-const maxCountOptions = [20, 50, 100];
+import { TaskRow } from "@/components/taskRow";
 
 const Page = () => {
+  const maxCountOptions = [20, 50, 100];
   const [maxCount, setMaxCount] = useState(20);
   const [tasks, setTasks] = useState<Task[]>([]);
   const [pageInfo, setPageInfo] = useState<PageInfo | null>(null);
   const endPage = pageInfo && Math.ceil(pageInfo.totalCount / pageInfo.limit);
-  const { projects } = useContext(ProjectsContext);
-  const [openProjectTaskId, setProjectOpenTaskId] = useState<string | null>(
-    null
-  );
-  const [openStatusTaskId, setStatusOpenTaskId] = useState<string | null>(
-    null
-  );
-  const handleDropdownClick = (taskId: string, column: string) => {
-    switch (column) {
-      case "project":
-        setProjectOpenTaskId(openProjectTaskId === taskId ? null : taskId); // 既に開いている場合は閉じ、閉じている場合は開く
-        break;
-      case "status":
-        setStatusOpenTaskId(openStatusTaskId === taskId ? null : taskId);
-        break;
-    }
-  };
-  const handleProjectSelect = (taskId: string, projectId: string) => {
-    const newProject = projects.find((project) => project.id === projectId);
-    // 見つからなかった場合は処理を中断
-    if (!newProject) {
-      alert("選択したプロジェクトが存在しません");
-      return;
-    }
-    setTasks((prevTasks) =>
-      prevTasks.map((task) =>
-        task.id === taskId
-          ? {
-              ...task,
-              project: newProject,
-            }
-          : task
-      )
-    );
-  };
-  const handleStatusSelect = (taskId: string, status: string) => {
-    setTasks((prevTasks) =>
-      prevTasks.map((task) =>
-        task.id === taskId
-          ? {
-              ...task,
-              status,
-            }
-          : task
-      )
-    );
-  };
   const handleMaxCount = async (count: number) => {
     try {
       setMaxCount(count);
@@ -158,91 +101,7 @@ const Page = () => {
           </div>
           <div>
             {tasks.map((task) => {
-              return (
-                <div key={task.id} className={styles.tableRow}>
-                  <div
-                    className={`${styles.tableCell} ${styles.tableCellTask}`}
-                  >
-                    <div className={styles.taskContent}>{task.title}</div>
-                  </div>
-                  <div
-                    className={`${styles.tableCell} ${styles.tableCellProject}`}
-                    onClick={() => handleDropdownClick(task.id, "project")}
-                  >
-                    <div
-                      className={`${styles.tableSelector} ${styles.selectContainer}`}
-                    >
-                      <div className={styles.selectValueContainer}>
-                        <p className={styles.projectName}>
-                          {task.project.name}
-                        </p>
-                        <div className={styles.iconContainer}>
-                          <FontAwesomeIcon icon={faChevronDown} />
-                        </div>
-                      </div>
-                      <div className={styles.selectPullDownShow}>
-                        {openProjectTaskId === task.id && (
-                          <ul className={styles.selectPullDown}>
-                            {projects.map((project) => (
-                              <li
-                                key={project.id}
-                                onClick={() =>
-                                  handleProjectSelect(task.id, project.id)
-                                }
-                                className={styles.selectOption}
-                              >
-                                {project.name}
-                              </li>
-                            ))}
-                          </ul>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                  <div
-                    className={`${styles.tableCell} ${styles.tableCellStatus}`}
-                    onClick={() => handleDropdownClick(task.id, "status")}
-                  >
-                    <div
-                      className={`${styles.tableSelector} ${styles.selectContainer}`}
-                    >
-                      <div className={styles.selectValueContainer}>
-                        <p className={styles.projectName}>
-                          {getStatusLabel(task.status)}
-                        </p>
-                        <div className={styles.iconContainer}>
-                          <FontAwesomeIcon icon={faChevronDown} />
-                        </div>
-                      </div>
-                      <div className={styles.selectPullDownShow}>
-                        {openStatusTaskId === task.id && (
-                          <ul className={styles.selectPullDown}>
-                            {Object.entries(statusMap).map(([key, value]) => (
-                              <li
-                                key={key}
-                                onClick={() => handleStatusSelect(task.id, key)}
-                                className={styles.selectOption}
-                              >
-                                {value}
-                              </li>
-                            ))}
-                          </ul>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                  <div
-                    className={`${styles.tableCell} ${styles.tableCellDeadline}`}
-                  >
-                    {dayjs(task.deadline).format("YYYY/MM/DD")}
-                  </div>
-                  <div
-                    className={`${styles.tableCell} ${styles.tableCellDetail}`}
-                  >
-                    <IoArrowForwardOutline size={12} />
-                  </div>
-                </div>
-              );
+              return <TaskRow key={task.id} task={task} setTasks={setTasks} />;
             })}
           </div>
         </div>
